requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b2
click==8.0.1
dj-static==0.0.6
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Markdown==3.3.4
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
